Refactor: Extract repository interfaces to data-api module
This commit extracts repository interfaces from the `core:data` module to a new `core:data
-api` module.

This change improves modularity and separation of concerns by isolating the repository interfaces from their implementations. It also allows other modules to depend on the `core:data-api` module without having to depend on the entire `core:data` module.

The following repository interfaces have been moved
:

- `CacheRepository`
- `FirebaseRepository`
- `NotificationRepository`
- `PromptRepository`
- `PreferenceManager`

The corresponding implementations in the `core:data` module have been updated to implement these interfaces.
