+++ app-arducon/src/main/java/com/keelim/arducon/ui/screen/main/MainViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.keelim.core.data.source.ArduconRepository
+import com.keelim.core.database.repository.ArduconRepository
 import com.keelim.core.database.model.DeepLink
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
