+++ app-my-grade/src/main/java/com/keelim/mygrade/ui/screen/grade/notes/NotesViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.keelim.commonAndroid.model.SealedUiState
 import com.keelim.commonAndroid.model.asSealedUiState
-import com.keelim.core.data.source.note.NoteRepository
+import com.keelim.core.database.repository.NoteRepository
 import com.keelim.core.database.model.Notices
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.SharingStarted
