+++ app-my-grade/src/main/java/com/keelim/mygrade/ui/screen/history/HistoryViewModel.kt
@@ -3,7 +3,7 @@ package com.keelim.mygrade.ui.screen.history
 import androidx.compose.runtime.Stable
 import androidx.lifecycle.ViewModel
 import com.keelim.common.di.DefaultDispatcher
-import com.keelim.core.data.source.HistoryRepository
+import com.keelim.core.database.repository.HistoryRepository
 import com.keelim.core.database.model.SimpleHistory
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.collections.immutable.PersistentList
