+++ app-my-grade/src/main/java/com/keelim/mygrade/ui/screen/task/TaskViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.keelim.commonAndroid.model.SealedUiState
 import com.keelim.commonAndroid.model.asSealedUiState
-import com.keelim.core.data.source.DefaultTaskRepository
+import com.keelim.core.database.repository.DefaultTaskRepository
 import com.keelim.core.database.model.LocalTask
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.SharingStarted
