+++ app-my-grade/src/main/java/com/keelim/mygrade/ui/screen/timer/history/TimerHistoryViewModel.kt
@@ -2,7 +2,7 @@ package com.keelim.mygrade.ui.screen.timer.history
 
 import androidx.compose.runtime.Stable
 import androidx.lifecycle.ViewModel
-import com.keelim.core.data.source.HistoryRepository
+import com.keelim.core.database.repository.HistoryRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 @Stable
