+++ core/data-api/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    alias(libs.plugins.keelim.jvm.library)
+}
+
+dependencies {
+    implementation(projects.core.model)
+    implementation(libs.kotlinx.coroutines.core)
+}
