+++ core/data-api/src/main/java/com/keelim/data/repository/CacheRepository.kt
@@ -1,4 +1,4 @@
-package com.keelim.core.data.source.cache
+package com.keelim.data.repository
 
 interface CacheRepository {
     suspend fun getResponse(url: String, enforce: Boolean = false): String
