+++ core/data-api/src/main/java/com/keelim/data/repository/NotificationRepository.kt
@@ -1,0 +1,5 @@
+package com.keelim.data.repository
+
+interface NotificationRepository {
+    suspend fun getNotification(): List<Notification>
+}
