+++ core/data-api/src/main/java/com/keelim/data/repository/PromptRepository.kt
@@ -1,4 +1,4 @@
-package com.keelim.core.data.source.prompt
+package com.keelim.data.repository
 
 interface PromptRepository {
     suspend fun getContent(prompt: String): Result<String>
