+++ core/data/build.gradle.kts
@@ -13,6 +13,7 @@ android { namespace = "com.keelim.core.data" }
 dependencies {
     api(projects.core.common)
     api(projects.core.database)
+    api(projects.core.dataApi)
     api(projects.core.model)
     api(projects.core.network)
