+++ core/data/src/main/java/com/keelim/core/data/di/RepositoryModule.kt
@@ -1,20 +1,20 @@
 package com.keelim.core.data.di
 
-import com.keelim.core.data.source.ArduconRepository
+import com.keelim.core.database.repository.ArduconRepository
 import com.keelim.core.data.source.ArduconRepositoryImpl
-import com.keelim.core.data.source.DefaultTaskRepository
+import com.keelim.core.database.repository.DefaultTaskRepository
 import com.keelim.core.data.source.DefaultTaskRepositoryImpl
-import com.keelim.core.data.source.HistoryRepository
+import com.keelim.core.database.repository.HistoryRepository
 import com.keelim.core.data.source.HistoryRepositoryImpl
-import com.keelim.core.data.source.NandaIORepository
+import com.keelim.core.database.repository.NandaIORepository
 import com.keelim.core.data.source.NandaIORepositoryImpl
-import com.keelim.core.data.source.firebase.FirebaseRepository
+import com.keelim.data.repository.FirebaseRepository
 import com.keelim.core.data.source.firebase.FirebaseRepositoryImpl
-import com.keelim.core.data.source.note.NoteRepository
+import com.keelim.core.database.repository.NoteRepository
 import com.keelim.core.data.source.note.NoteRepositoryImpl
-import com.keelim.core.data.source.notification.NotificationRepository
+import com.keelim.data.repository.NotificationRepository
 import com.keelim.core.data.source.notification.NotificationRepositoryImpl
-import com.keelim.core.data.source.prompt.PromptRepository
+import com.keelim.data.repository.PromptRepository
 import com.keelim.core.data.source.prompt.PromptRepositoryImpl
 import dagger.Binds
 import dagger.Module
