+++ core/data/src/main/java/com/keelim/core/data/source/local/SharedPreferenceManager.kt
@@ -18,6 +18,7 @@ package com.keelim.core.data.source.local
 import android.app.Activity
 import android.content.Context
 import androidx.core.content.edit
+import com.keelim.data.repository.PreferenceManager
 import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
