+++ core/data/src/main/java/com/keelim/core/data/source/note/NoteRepositoryImpl.kt
@@ -3,6 +3,7 @@ package com.keelim.core.data.source.note
 import com.keelim.common.di.IoDispatcher
 import com.keelim.core.database.dao.NoteDao
 import com.keelim.core.database.model.Notices
+import com.keelim.core.database.repository.NoteRepository
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.mapLatest
