--- core/data/src/main/java/com/keelim/core/data/source/notification/NotificationRepository.kt
@@ -1,12 +1,0 @@
-package com.keelim.core.data.source.notification
-
-data class Notification(
-    val date: String,
-    val title: String,
-    val desc: String,
-    val fixed: Boolean,
-)
-
-interface NotificationRepository {
-    suspend fun getNotification(): List<Notification>
-}
