+++ core/database/src/main/java/com/keelim/core/database/repository/ArduconDataSource.kt
@@ -1,9 +1,9 @@
-package com.keelim.core.data.source
+package com.keelim.core.database.repository
 
 import com.keelim.core.database.model.DeepLink
 import kotlinx.coroutines.flow.Flow
 
-interface ArduconRepository {
+interface ArduconDataSource {
     suspend fun insertDeepLinkUrl(deepLink: DeepLink)
     fun getDeepLinkUrls(): Flow<List<DeepLink>>
     suspend fun deleteDeepLinkUrl(deepLink: DeepLink)
