+++ core/database/src/main/java/com/keelim/core/database/repository/ArduconRepository.kt
@@ -1,9 +1,9 @@
-package com.keelim.core.data.source.local
+package com.keelim.core.database.repository
 
 import com.keelim.core.database.model.DeepLink
 import kotlinx.coroutines.flow.Flow
 
-interface ArduconDataSource {
+interface ArduconRepository {
     suspend fun insertDeepLinkUrl(deepLink: DeepLink)
     fun getDeepLinkUrls(): Flow<List<DeepLink>>
     suspend fun deleteDeepLinkUrl(deepLink: DeepLink)
