+++ core/database/src/main/java/com/keelim/core/database/repository/DefaultTaskRepository.kt
@@ -1,4 +1,4 @@
-package com.keelim.core.data.source
+package com.keelim.core.database.repository
 
 import com.keelim.core.database.model.LocalTask
 import kotlinx.coroutines.flow.Flow
