+++ core/model/src/main/java/com/keelim/model/Notification.kt
@@ -1,0 +1,8 @@
+package com.keelim.data.repository
+
+data class Notification(
+    val date: String,
+    val title: String,
+    val desc: String,
+    val fixed: Boolean,
+)
